requirements.txt
pyautogui
pillow
opencv-python
openai
python-dotenv
mss
numpy
